I separated all components in containers and presentational based on this rules https://redux.js.org/basics/usage-with-react#presentational-and-container-components
