+++ src/components/containers/Credits.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import * as S from './Presentational/App.styles';
+import * as S from '../presentationals/App.styles';
 import Filter from './Filter';
 import Grid from './Grid';
-import { selectCredits } from '../slices/home/homeSlice';
+import { selectCredits } from '../../slices/home/homeSlice';
 
 const Credits = () => {
   const credits = useSelector(selectCredits);
