+++ src/components/containers/Grid.js
@@ -2,8 +2,8 @@ import React, { useEffect } from 'react';
 import PropTypes, { object } from 'prop-types';
 import { useSelector, useDispatch } from 'react-redux';
 import { useParams } from 'react-router-dom';
-import * as S from '../Presentational/Grid.styles';
-import GridItem from './GridItem';
+import * as S from '../presentationals/Grid.styles';
+import GridItem from '../presentationals/GridItem';
 import { selectFilters, updateFilter } from '../../slices/current/currentSlice';
 
 const Grid = ({ items, url }) => {
@@ -11,7 +11,7 @@ const Grid = ({ items, url }) => {
   const dispatch = useDispatch();
   useEffect(() => {
     dispatch(updateFilter({ name: '', selectFilter: { type: category, state: 'All' } }));
-  }, [dispatch, category]);
+  }, []); // eslint-disable-line
   const gridStyles = [
     'pattern-checks-md',
     'pattern-diagonal-lines-md',
