+++ src/components/containers/HomePage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import Grid from './Grid';
-import { selectCategories } from '../slices/home/homeSlice';
+import { selectCategories } from '../../slices/home/homeSlice';
 
 const HomePage = () => {
   const categories = useSelector(selectCategories);
