+++ src/components/containers/Item.js
@@ -1,11 +1,11 @@
 import React, { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
-import Filter from '../Filter';
+import Filter from './Filter';
 import { selectInfo, updateFromApi, selectLanguage } from '../../slices/current/currentSlice';
-import * as S from '../Presentational/Item.styles';
-import Attributes from './Attributes';
-import Title from './Title';
+import * as S from '../presentationals/Item.styles';
+import Attributes from '../presentationals/Attributes';
+import Title from '../presentationals/Title';
 
 const Item = () => {
   const { category, item } = useParams();
@@ -15,8 +15,7 @@ const Item = () => {
 
   useEffect(() => {
     if (!Object.keys(dataArray).length) dispatch(updateFromApi(category.toLowerCase(), language));
-    // window.scrollTo(0, 0);
-  }, [category, dispatch, dataArray, language]);
+  }, []); // eslint-disable-line
 
   const currentData = Object.values(dataArray).filter(data => data.name[`name-${language}`] === item)[0];
   return (
