+++ src/components/presentationals/Loader.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import * as S from './Presentational/App.styles';
-import logo from '../images/logo.svg';
+import * as S from './App.styles';
+import logo from '../../images/logo.svg';
 
 const Loader = ({ hidden }) => (
   <S.Loader hidden={hidden}>
