+++ src/index.js
@@ -1,7 +1,7 @@
 import './index.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App';
+import App from './components/containers/App';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
