+++ src/tests/Grid/Grid.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Grid from '../../components/Grid';
+import Grid from '../../components/containers/Grid';
 import store from '../../store';
 import '@testing-library/jest-dom/extend-expect';
