+++ src/tests/unit/openCategory.test.js
@@ -3,7 +3,7 @@ import {
 } from '@testing-library/react';
 import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
-import App from '../../components/App';
+import App from '../../components/containers/App';
 import '../globals';
 
 afterAll(() => {
